version for open genshin 4.7 patch
original version encrypts the config file for reasons unknown and also either fails to read it or fails to save to it on my PC so I changed initial values to my preference. I replaced the config file with the reshade ini file code with some changes.... which doesn't work so you just have to open the overlay with insert. Will fix soon or just publish a build with preset values. Was also working on adding it as a reshade addon just to consolidate the UI as an addon overlay but that fork got obliterated